+++ models.go
@@ -1,36 +1,23 @@
 package gpt3
 
-import "fmt"
+import (
+	"context"
+	"fmt"
+)
 
-// APIError represents an error that occurred on an API
-type APIError struct {
-	StatusCode int    `json:"status_code"`
-	Message    string `json:"message"`
-	Type       string `json:"type"`
-}
-
-func (e APIError) Error() string {
-	return fmt.Sprintf("[%d:%s] %s", e.StatusCode, e.Type, e.Message)
-}
-
-// APIErrorResponse is the full error response that has been returned by an API.
-type APIErrorResponse struct {
-	Error APIError `json:"error"`
-}
-
-// EngineObject contained in an engine response
-type EngineObject struct {
-	ID     string `json:"id"`
-	Object string `json:"object"`
-	Owner  string `json:"owner"`
-	Ready  bool   `json:"ready"`
-}
-
-// EnginesResponse is returned from the Engines API
-type EnginesResponse struct {
-	Data   []EngineObject `json:"data"`
-	Object string         `json:"object"`
-}
+// Model Types
+const (
+	TextAda001Model     = "text-ada-001"
+	TextBabbage001Model = "text-babbage-001"
+	TextCurie001Model   = "text-curie-001"
+	TextDavinci001Model = "text-davinci-001"
+	TextDavinci002Model = "text-davinci-002"
+	TextDavinci003Model = "text-davinci-003"
+	AdaModel            = "ada"
+	BabbageModel        = "babbage"
+	CurieModel          = "curie"
+	DavinciModel        = "davinci"
+)
 
 // ModelObject
 type ModelObject struct {
@@ -46,306 +33,40 @@ type ModelsResponse struct {
 	Object string        `json:"object"`
 }
 
-// CompletionRequest is a request for the completions API
-type CompletionRequest struct {
-	// A list of string prompts to use.
-	// TODO there are other prompt types here for using token integers that we could add support for.
-	Prompt []string `json:"prompt"`
-	// How many tokens to complete up to. Max of 512
-	MaxTokens *int `json:"max_tokens,omitempty"`
-	// Sampling temperature to use
-	Temperature *float32 `json:"temperature,omitempty"`
-	// Alternative to temperature for nucleus sampling
-	TopP *float32 `json:"top_p,omitempty"`
-	// How many choice to create for each prompt
-	N *int `json:"n"`
-	// Include the probabilities of most likely tokens
-	LogProbs *int `json:"logprobs"`
-	// Echo back the prompt in addition to the completion
-	Echo bool `json:"echo"`
-	// Up to 4 sequences where the API will stop generating tokens. Response will not contain the stop sequence.
-	Stop []string `json:"stop,omitempty"`
-	// PresencePenalty number between 0 and 1 that penalizes tokens that have already appeared in the text so far.
-	PresencePenalty float32 `json:"presence_penalty"`
-	// FrequencyPenalty number between 0 and 1 that penalizes tokens on existing frequency in the text so far.
-	FrequencyPenalty float32 `json:"frequency_penalty"`
+// Models lists the currently available models, and provides basic information about each
+// option such as the owner and permissioning.
+func (c *client) Models(ctx context.Context) (*ModelsResponse, error) {
+	req, err := c.newRequest(ctx, "GET", "/models", nil)
+	if err != nil {
+		return nil, err
+	}
+	resp, err := c.performRequest(req)
+	if err != nil {
+		return nil, err
+	}
 
-	// Whether to stream back results or not. Don't set this value in the request yourself
-	// as it will be overridden depending on if you use CompletionStream or Completion methods.
-	Stream bool `json:"stream,omitempty"`
+	output := new(ModelsResponse)
+	if err := getResponseObject(resp, output); err != nil {
+		return nil, err
+	}
+	return output, nil
 }
 
-// EditsRequest is a request for the edits API
-type EditsRequest struct {
-	// ID of the model to use. You can use the List models API to see all of your available models, or see our Model overview for descriptions of them.
-	Model string `json:"model"`
-	// The input text to use as a starting point for the edit.
-	Input string `json:"input"`
-	// The instruction that tells the model how to edit the prompt.
-	Instruction string `json:"instruction"`
-	// Sampling temperature to use
-	Temperature *float32 `json:"temperature,omitempty"`
-	// Alternative to temperature for nucleus sampling
-	TopP *float32 `json:"top_p,omitempty"`
-	// How many edits to generate for the input and instruction. Defaults to 1
-	N *int `json:"n"`
-}
+// Model retrieves a single model, providing basic information about the model such
+// as the owner and permissioning.
+func (c *client) Model(ctx context.Context, model string) (*ModelObject, error) {
+	req, err := c.newRequest(ctx, "GET", fmt.Sprintf("/models/%s", model), nil)
+	if err != nil {
+		return nil, err
+	}
+	resp, err := c.performRequest(req)
+	if err != nil {
+		return nil, err
+	}
 
-// EmbeddingsRequest is a request for the Embeddings API
-type EmbeddingsRequest struct {
-	// Input text to get embeddings for, encoded as a string or array of tokens. To get embeddings
-	// for multiple inputs in a single request, pass an array of strings or array of token arrays.
-	// Each input must not exceed 2048 tokens in length.
-	Input []string `json:"input"`
-	// ID of the model to use
-	Model string `json:"model"`
-	// The request user is an optional parameter meant to be used to trace abusive requests
-	// back to the originating user. OpenAI states:
-	// "The [user] IDs should be a string that uniquely identifies each user. We recommend hashing
-	// their username or email address, in order to avoid sending us any identifying information.
-	// If you offer a preview of your product to non-logged in users, you can send a session ID
-	// instead."
-	User string `json:"user,omitempty"`
-}
-
-// UploadFileRequest is a request for the Files API
-type UploadFileRequest struct {
-	// The file name of the JSON Lines file to upload
-	File string `json:"file"`
-	// The purpose of the file. Use "fine-tune" for a file that will be used to fine-tune a model.
-	Purpose string `json:"purpose"`
-}
-
-// CreateFineTuneRequest is a request for the FineTune API
-type CreateFineTuneRequest struct {
-	// The ID of an uploaded file that contains training data
-	TrainingFile string `json:"training_file"`
-	// The ID of an uploaded file that contains validation data
-	ValidationFile string `json:"validation_file"`
-	// The ID of the model to fine-tune
-	Model string `json:"model"`
-	// The number of epochs to train for
-	NEpochs int `json:"n_epochs"`
-	// The batch size to use for training
-	BatchSize int `json:"batch_size"`
-	// The learning rate to use for training
-	LearningRate float32 `json:"learning_rate"`
-	// The weight to use for loss on the prompt tokens
-	PromptLossWeight float32 `json:"prompt_loss_weight"`
-	// If set, we calculate classification-specific metrics using the validation set at the end of each epoch
-	ComputeClassificationMetrics bool `json:"compute_classification_metrics"`
-	// The number of classes in a classification task
-	ClassificationNClasses int `json:"classification_n_classes"`
-	// The positive class in binary classification
-	ClassificationPositiveClass string `json:"classification_positive_class"`
-	// If this is provided, we calculate F-beta scores at the specified beta values
-	ClassificationBetas []float32 `json:"classification_betas"`
-	// A string of up to 40 characters that will be added to your fine-tuned model name
-	Suffix string `json:"suffix"`
-}
-
-// FineTuneRequest is a request for the FineTune API
-type FineTuneRequest struct {
-	// The ID of the fine-tune job
-	FineTuneID string `json:"fine_tune_id"`
-}
-
-// FineTuneEventsRequest is a request for the FineTune API
-type FineTuneEventsRequest struct {
-	// The ID of the fine-tune job
-	FineTuneID string `json:"fine_tune_id"`
-	// Whether to stream events for the fine-tune job
-	Stream bool `json:"stream"`
-}
-
-// DeleteFineTuneModelRequest is a request for the FineTune API
-type DeleteFineTuneModelRequest struct {
-	// The ID of the fine-tune model to delete
-	Model string `json:"model"`
-}
-
-// LogprobResult represents logprob result of Choice
-type LogprobResult struct {
-	Tokens        []string             `json:"tokens"`
-	TokenLogprobs []float32            `json:"token_logprobs"`
-	TopLogprobs   []map[string]float32 `json:"top_logprobs"`
-	TextOffset    []int                `json:"text_offset"`
-}
-
-// CompletionResponseChoice is one of the choices returned in the response to the Completions API
-type CompletionResponseChoice struct {
-	Text         string        `json:"text"`
-	Index        int           `json:"index"`
-	LogProbs     LogprobResult `json:"logprobs"`
-	FinishReason string        `json:"finish_reason"`
-}
-
-// CompletionResponse is the full response from a request to the completions API
-type CompletionResponse struct {
-	ID      string                     `json:"id"`
-	Object  string                     `json:"object"`
-	Created int                        `json:"created"`
-	Model   string                     `json:"model"`
-	Choices []CompletionResponseChoice `json:"choices"`
-	Usage   CompletionResponseUsage    `json:"usage"`
-}
-
-// CompletionResponseUsage is the object that returns how many tokens the completion's request used
-type CompletionResponseUsage struct {
-	PromptTokens     int `json:"prompt_tokens"`
-	CompletionTokens int `json:"completion_tokens"`
-	TotalTokens      int `json:"total_tokens"`
-}
-
-// EditsResponse is the full response from a request to the edits API
-type EditsResponse struct {
-	Object  string                `json:"object"`
-	Created int                   `json:"created"`
-	Choices []EditsResponseChoice `json:"choices"`
-	Usage   EditsResponseUsage    `json:"usage"`
-}
-
-// The inner result of a create embeddings request, containing the embeddings for a single input.
-type EmbeddingsResult struct {
-	// The type of object returned (e.g., "list", "object")
-	Object string `json:"object"`
-	// The embedding data for the input
-	Embedding []float64 `json:"embedding"`
-	Index     int       `json:"index"`
-}
-
-// The usage stats for an embeddings response
-type EmbeddingsUsage struct {
-	// The number of tokens used by the prompt
-	PromptTokens int `json:"prompt_tokens"`
-	// The total tokens used
-	TotalTokens int `json:"total_tokens"`
-}
-
-// EmbeddingsResponse is the response from a create embeddings request.
-//
-// See: https://beta.openai.com/docs/api-reference/embeddings/create
-type EmbeddingsResponse struct {
-	Object string             `json:"object"`
-	Data   []EmbeddingsResult `json:"data"`
-	Usage  EmbeddingsUsage    `json:"usage"`
-}
-
-// FileObject is a single file object
-type FileObject struct {
-	ID        string `json:"id"`
-	Object    string `json:"object"`
-	Bytes     int    `json:"bytes"`
-	CreatedAt int    `json:"created_at"`
-	Filename  string `json:"filename"`
-	Purpose   string `json:"purpose"`
-}
-
-// FilesResponse is the response from a list files request.
-//
-// See: https://beta.openai.com/docs/api-reference/files/list
-type FilesResponse struct {
-	Data   []FileObject `json:"data"`
-	Object string       `json:"object"`
-}
-
-// DeleteFileResponse is the response from a delete file request.
-//
-// See: https://beta.openai.com/docs/api-reference/files/delete
-type DeleteFileResponse struct {
-	ID      string `json:"id"`
-	Object  string `json:"object"`
-	Deleted bool   `json:"deleted"`
-}
-
-// FineTuneEvent is a single fine tune event
-type FineTuneEvent struct {
-	Object    string `json:"object"`
-	CreatedAt int    `json:"created_at"`
-	Level     string `json:"level"`
-	Message   string `json:"message"`
-}
-
-// FineTuneHyperparams is the hyperparams for a fine tune request
-type FineTuneHyperparams struct {
-	BatchSize              int     `json:"batch_size"`
-	LearningRateMultiplier float64 `json:"learning_rate_multiplier"`
-	NEpochs                int     `json:"n_epochs"`
-	PromptLossWeight       float64 `json:"prompt_loss_weight"`
-}
-
-// FineTuneObject is a single fine tune object
-//
-// See: https://beta.openai.com/docs/api-reference/fine-tunes/retrieve
-type FineTuneObject struct {
-	ID              string              `json:"id"`
-	Object          string              `json:"object"`
-	Model           string              `json:"model"`
-	CreatedAt       int                 `json:"created_at"`
-	Events          []FineTuneEvent     `json:"events"`
-	FineTuneModel   string              `json:"fine_tune_model"`
-	Hyperparams     FineTuneHyperparams `json:"hyperparams"`
-	OrganizationID  string              `json:"organization_id"`
-	ResultFiles     []FileObject        `json:"result_files"`
-	Status          string              `json:"status"`
-	ValidationFiles []FileObject        `json:"validation_files"`
-	TrainingFiles   []FileObject        `json:"training_files"`
-	UpdatedAt       int                 `json:"updated_at"`
-}
-
-// FineTunesResponse is the response from a list fine tunes request.
-//
-// See: https://beta.openai.com/docs/api-reference/fine-tunes/list
-type FineTunesResponse struct {
-	Data   []FineTuneObject `json:"data"`
-	Object string           `json:"object"`
-}
-
-// FineTuneEventsResponse
-//
-// See: https://beta.openai.com/docs/api-reference/fine-tunes/events
-type FineTuneEventsResponse struct {
-	Data   []FineTuneEvent `json:"data"`
-	Object string          `json:"object"`
-}
-
-// DeleteFineTuneModelResponse
-//
-// See: https://beta.openai.com/docs/api-reference/fine-tunes/delete-model
-type DeleteFineTuneModelResponse struct {
-	ID      string `json:"id"`
-	Object  string `json:"object"`
-	Deleted bool   `json:"deleted"`
-}
-
-// EditsResponseChoice is one of the choices returned in the response to the Edits API
-type EditsResponseChoice struct {
-	Text  string `json:"text"`
-	Index int    `json:"index"`
-}
-
-// EditsResponseUsage is a structure used in the response from a request to the edits API
-type EditsResponseUsage struct {
-	PromptTokens     int `json:"prompt_tokens"`
-	CompletionTokens int `json:"completion_tokens"`
-	TotalTokens      int `json:"total_tokens"`
-}
-
-// SearchRequest is a request for the document search API
-type SearchRequest struct {
-	Documents []string `json:"documents"`
-	Query     string   `json:"query"`
-}
-
-// SearchData is a single search result from the document search API
-type SearchData struct {
-	Document int     `json:"document"`
-	Object   string  `json:"object"`
-	Score    float64 `json:"score"`
-}
-
-// SearchResponse is the full response from a request to the document search API
-type SearchResponse struct {
-	Data   []SearchData `json:"data"`
-	Object string       `json:"object"`
+	output := new(ModelObject)
+	if err := getResponseObject(resp, output); err != nil {
+		return nil, err
+	}
+	return output, nil
 }
